requirements.txt
Django>=3.1.0,<3.1.9
djangorestframework>=3.11.1,<3.11.9
psycopg2>=2.8.5,<2.8.9
drf-nested-routers==0.91